Split up common_xml_util into common_functions and converters
