+++ masci_tools/io/parsers/fleur/fleur_schema/outschema_todict.py
@@ -15,7 +15,7 @@ This module provides the functionality to create/load the schema_dict for the
 FleurInputSchema.xsd
 """
 from .fleur_schema_parser_functions import *  #pylint: disable=unused-wildcard-import
-from masci_tools.util.xml.common_xml_util import clear_xml
+from masci_tools.util.xml.common_functions import clear_xml
 from masci_tools.util.case_insensitive_dict import CaseInsensitiveDict
 from lxml import etree
 import copy
