+++ masci_tools/tests/test_xml_converters.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-Test of the functions in common_xml_util
+Test of the functions in masci_tools.util.xml.converters
 """
 import pytest
 import os
@@ -13,12 +13,11 @@ CLEAR_XML_TEST_FILE = os.path.abspath(os.path.join(TEST_FOLDER, 'files/fleur/tes
 
 LOGGER = logging.getLogger(__name__)
 
-
 def test_convert_from_fortran_bool():
     """
     Test of the convert_from_fortran_bool function
     """
-    from masci_tools.util.xml.common_xml_util import convert_from_fortran_bool
+    from masci_tools.util.xml.converters import convert_from_fortran_bool
 
     TRUE_ITEMS = ('T', 't', True)
     FALSE_ITEMS = ('F', 'f', False)
@@ -40,7 +39,7 @@ def test_convert_to_fortran_bool():
     """
     Test of the function convert_to_fortran_bool
     """
-    from masci_tools.util.xml.common_xml_util import convert_to_fortran_bool
+    from masci_tools.util.xml.converters import convert_to_fortran_bool
 
     TRUE_ITEMS = (True, 'True', 't', 'T')
     FALSE_ITEMS = (False, 'False', 'f', 'F')
@@ -56,124 +55,6 @@ def test_convert_to_fortran_bool():
 
     with pytest.raises(TypeError, match='convert_to_fortran_bool accepts only a string or bool as argument'):
         convert_to_fortran_bool(())
-
-
-def test_eval_xpath(caplog):
-    """
-    Test of the eval_xpath function
-    """
-    from lxml import etree
-    from masci_tools.util.xml.common_xml_util import eval_xpath
-
-    parser = etree.XMLParser(attribute_defaults=True, encoding='utf-8')
-    xmltree = etree.parse(CLEAR_XML_TEST_FILE, parser)
-    root = xmltree.getroot()
-
-    scfLoop = eval_xpath(root, '//scfLoop')
-    assert isinstance(scfLoop, etree._Element)
-
-    scfLoop = eval_xpath(root, '//scfLoop', list_return=True)
-    assert len(scfLoop) == 1
-    assert isinstance(scfLoop[0], etree._Element)
-
-    include_tags = eval_xpath(root,
-                              '//xi:include',
-                              namespaces={'xi': 'http://www.w3.org/2001/XInclude'},
-                              list_return=True)
-    assert len(include_tags) == 2
-    assert isinstance(include_tags[0], etree._Element)
-
-    species_z = eval_xpath(root, "//species[@name='Cu-1']/@atomicNumber")
-    assert species_z == '29'
-
-    ldau_tags = eval_xpath(root, "//species[@name='Cu-1']/ldaU")
-    assert ldau_tags == []
-
-    with pytest.raises(ValueError, match='There was a XpathEvalError on the xpath:'):
-        ldau_tags = eval_xpath(root, "//species/[@name='Cu-1']/ldaU")
-
-    with caplog.at_level(logging.WARNING):
-        with pytest.raises(ValueError, match='There was a XpathEvalError on the xpath:'):
-            ldau_tags = eval_xpath(root, "//species/[@name='Cu-1']/ldaU", logger=LOGGER)
-
-    assert 'There was a XpathEvalError on the xpath:' in caplog.text
-
-
-def test_clear_xml():
-    """
-    Test of the clear_xml function
-    """
-    from lxml import etree
-    from masci_tools.util.xml.common_xml_util import eval_xpath, clear_xml
-    parser = etree.XMLParser(attribute_defaults=True, encoding='utf-8')
-    xmltree = etree.parse(CLEAR_XML_TEST_FILE, parser)
-
-    #Check that the file contains comments and includes
-    root = xmltree.getroot()
-    comments = eval_xpath(root, '//comment()', list_return=True)
-    assert len(comments) == 3
-
-    include_tags = eval_xpath(root,
-                              '//xi:include',
-                              namespaces={'xi': 'http://www.w3.org/2001/XInclude'},
-                              list_return=True)
-    assert len(include_tags) == 2
-
-    symmetry_tags = eval_xpath(root, '//symOp', list_return=True)
-    assert len(symmetry_tags) == 0
-
-    cleared_tree = clear_xml(xmltree)
-    cleared_root = cleared_tree.getroot()
-    old_root = xmltree.getroot()
-
-    #Make sure that the original tree was not modified
-    comments = eval_xpath(old_root, '//comment()', list_return=True)
-    assert len(comments) == 3
-
-    #Check that the cleared tree is correct
-    comments = eval_xpath(cleared_root, '//comment()', list_return=True)
-    assert len(comments) == 0
-
-    include_tags = eval_xpath(cleared_root,
-                              '//xi:include',
-                              namespaces={'xi': 'http://www.w3.org/2001/XInclude'},
-                              list_return=True)
-    assert len(include_tags) == 0
-
-    symmetry_tags = eval_xpath(cleared_root, '//symOp', list_return=True)
-    assert len(symmetry_tags) == 16
-
-
-def test_get_xml_attribute(caplog):
-    """
-    Test of the clear_xml function
-    """
-    from lxml import etree
-    from masci_tools.util.xml.common_xml_util import get_xml_attribute, eval_xpath
-    parser = etree.XMLParser(attribute_defaults=True, encoding='utf-8')
-    xmltree = etree.parse(CLEAR_XML_TEST_FILE, parser)
-    root = xmltree.getroot()
-
-    scfLoop = eval_xpath(root, '//scfLoop')
-    assert get_xml_attribute(scfLoop, 'alpha') == '.05000000'
-    assert get_xml_attribute(scfLoop, 'itmax') == '1'
-    assert get_xml_attribute(scfLoop, 'maxIterBroyd') == '99'
-
-    with pytest.raises(ValueError, match='Tried to get attribute: "TEST" from element scfLoop.'):
-        get_xml_attribute(scfLoop, 'TEST')
-
-    with caplog.at_level(logging.WARNING):
-        assert get_xml_attribute(scfLoop, 'TEST', logger=LOGGER) is None
-    assert 'Tried to get attribute: "TEST" from element scfLoop.' in caplog.text
-
-    with pytest.raises(TypeError,
-                       match='Can not get attributename: "TEST" from node of type <class \'lxml.etree._ElementTree\'>'):
-        get_xml_attribute(xmltree, 'TEST')
-
-    with caplog.at_level(logging.WARNING):
-        assert get_xml_attribute(xmltree, 'TEST', logger=LOGGER) is None
-    assert 'Can not get attributename: "TEST" from node of type <class \'lxml.etree._ElementTree\'>' in caplog.text
-
 
 TEST_STRINGS = ['1.2314', 'all', ['all', '213', '-12'], ['PI', 'NOT_PI', '1.2'], ['F', 'T'], ['F', 'T']]
 
@@ -194,7 +75,7 @@ def test_convert_xml_attribute(string_attr, types, results):
     """
     Test of the convert_xml_attribute function
     """
-    from masci_tools.util.xml.common_xml_util import convert_xml_attribute
+    from masci_tools.util.xml.converters import convert_xml_attribute
 
     expected_val, expected_suc = results
 
@@ -214,7 +95,7 @@ def test_convert_xml_attribute_warnings(caplog, string_attr, types, results, war
     """
     Test of the convert_xml_attribute function
     """
-    from masci_tools.util.xml.common_xml_util import convert_xml_attribute
+    from masci_tools.util.xml.converters import convert_xml_attribute
 
     expected_val, expected_suc = results
 
@@ -245,7 +126,7 @@ def test_convert_attribute_to_xml(attr_value, types, results):
     """
     Test of the convert_xml_attribute function
     """
-    from masci_tools.util.xml.common_xml_util import convert_attribute_to_xml
+    from masci_tools.util.xml.converters import convert_attribute_to_xml
 
     expected_val, expected_suc = results
     if not expected_suc:
@@ -263,7 +144,7 @@ def test_convert_attribute_to_xml_warnings(caplog, attr_value, types, results, w
     """
     Test of the convert_xml_attribute function
     """
-    from masci_tools.util.xml.common_xml_util import convert_attribute_to_xml
+    from masci_tools.util.xml.converters import convert_attribute_to_xml
 
     expected_val, expected_suc = results
 
@@ -336,7 +217,7 @@ def test_convert_xml_text(string_text, definitions, results):
     """
     Test of the convert_xml_attribute function
     """
-    from masci_tools.util.xml.common_xml_util import convert_xml_text
+    from masci_tools.util.xml.converters import convert_xml_text
 
     expected_val, expected_suc = results
 
@@ -355,7 +236,7 @@ def test_convert_xml_text_warnings(caplog, string_text, definitions, results, wa
     """
     Test of the convert_xml_attribute function
     """
-    from masci_tools.util.xml.common_xml_util import convert_xml_text
+    from masci_tools.util.xml.converters import convert_xml_text
 
     expected_val, expected_suc = results
 
@@ -390,7 +271,7 @@ def test_convert_text_to_xml(text_value, definitions, results):
     """
     Test of the convert_xml_attribute function
     """
-    from masci_tools.util.xml.common_xml_util import convert_text_to_xml
+    from masci_tools.util.xml.converters import convert_text_to_xml
 
     expected_val, expected_suc = results
     if not expected_suc:
@@ -408,7 +289,7 @@ def test_convert_text_to_xml_warnings(caplog, text_value, definitions, results, 
     """
     Test of the convert_xml_attribute function
     """
-    from masci_tools.util.xml.common_xml_util import convert_text_to_xml
+    from masci_tools.util.xml.converters import convert_text_to_xml
 
     expected_val, expected_suc = results
 
@@ -421,56 +302,3 @@ def test_convert_text_to_xml_warnings(caplog, text_value, definitions, results, 
     else:
         for expected_warning in warnings:
             assert expected_warning in caplog.text
-
-
-def test_split_off_tag():
-    """
-    Test of the split_off_tag function
-    """
-    from masci_tools.util.xml.common_xml_util import split_off_tag
-
-    assert split_off_tag('/fleurInput/calculationSetup/cutoffs') == ('/fleurInput/calculationSetup', 'cutoffs')
-    assert split_off_tag('/fleurInput/calculationSetup/cutoffs/') == ('/fleurInput/calculationSetup', 'cutoffs')
-    assert split_off_tag('./calculationSetup/cutoffs') == ('./calculationSetup', 'cutoffs')
-
-
-def test_split_off_attrib():
-    """
-    Test of the split_off_tag function
-    """
-    from masci_tools.util.xml.common_xml_util import split_off_attrib
-
-    assert split_off_attrib('/fleurInput/calculationSetup/cutoffs/@Kmax') == ('/fleurInput/calculationSetup/cutoffs',
-                                                                              'Kmax')
-    with pytest.raises(AssertionError):
-        split_off_attrib('/fleurInput/calculationSetup/cutoffs')
-    with pytest.raises(AssertionError):
-        split_off_attrib("/fleurInput/atomSpecies/species[@name='TEST']")
-    assert split_off_attrib('./calculationSetup/cutoffs/@Kmax') == ('./calculationSetup/cutoffs', 'Kmax')
-
-
-def test_check_complex_xpath(load_inpxml):
-    """
-    Test of the check_complex_xpath function
-    """
-    from masci_tools.util.xml.common_xml_util import check_complex_xpath
-
-    FILE_PATH = os.path.dirname(os.path.abspath(__file__))
-    TEST_INPXML_PATH = os.path.join(FILE_PATH, 'files/fleur/Max-R5/FePt_film_SSFT_LO/files/inp2.xml')
-
-    xmltree, _ = load_inpxml(TEST_INPXML_PATH)
-
-    check_complex_xpath(xmltree, '/fleurInput/atomSpecies/species', "/fleurInput/atomSpecies/species[@name='Fe-1']")
-
-    with pytest.raises(ValueError):
-        check_complex_xpath(xmltree, '/fleurInput/atomSpecies/species',
-                            "/fleurInput/atomSpecies/species[@name='Fe-1']/lo")
-
-    with pytest.raises(ValueError):
-        check_complex_xpath(xmltree, '/fleurInput/atomSpecies/species',
-                            "/fleurInput/atomSpecies/species[@name='Fe-1']/@name")
-
-    check_complex_xpath(xmltree, '/fleurInput/atomSpecies/species', '/fleurInput/atomSpecies/species')
-    check_complex_xpath(xmltree, '/fleurInput/atomSpecies/species',
-                        "/fleurInput/atomSpecies/species[@name='does_not_exist']")
-    check_complex_xpath(xmltree, '/fleurInput/atomSpecies/species/lo', "//species[@name='Pt-1']/lo")
