+++ masci_tools/tests/test_xml_setters_xpaths.py
@@ -12,7 +12,7 @@ TEST_INPXML_PATH = os.path.join(FILE_PATH, 'files/fleur/Max-R5/FePt_film_SSFT_LO
 
 def test_xml_create_tag_schema_dict(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_create_tag_schema_dict
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -40,7 +40,7 @@ def test_xml_create_tag_schema_dict(load_inpxml):
 
 def test_xml_create_tag_schema_dict_differing_xpaths(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_create_tag_schema_dict
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -67,7 +67,7 @@ def test_xml_create_tag_schema_dict_differing_xpaths(load_inpxml):
 
 def test_xml_create_tag_schema_dict_element(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_create_tag_schema_dict
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -89,7 +89,7 @@ def test_xml_create_tag_schema_dict_element(load_inpxml):
 
 def test_xml_create_tag_schema_dict_create_parents(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_create_tag_schema_dict
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -111,7 +111,7 @@ def test_xml_create_tag_schema_dict_create_parents(load_inpxml):
 
 def test_eval_xpath_create_existing(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import eval_xpath_create
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -127,7 +127,7 @@ def test_eval_xpath_create_existing(load_inpxml):
 
 def test_eval_xpath_create_non_existing(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import eval_xpath_create
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -144,7 +144,7 @@ def test_eval_xpath_create_non_existing(load_inpxml):
 
 def test_eval_xpath_create_differing_xpaths(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import eval_xpath_create
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -190,7 +190,7 @@ TEST_RESULTS = ['9.000000', '5.3210000000', 'Pi/4.0']
 @pytest.mark.parametrize('attribname, attribvalue, result', zip(TEST_ATTRIB_NAME, TEST_VALUES, TEST_RESULTS))
 def test_xml_set_attrib_value(load_inpxml, attribname, attribvalue, result):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_attrib_value
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -216,7 +216,7 @@ def test_xml_set_attrib_value_all(load_inpxml, attribv, expected_result, occurre
     Test of the functionality of xml_set_attrib_value_no_create with multiple occurrences
     of the sttribute and different values for occurrences
     """
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_attrib_value
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -240,7 +240,7 @@ def test_xml_set_attrib_value_differing_xpaths(load_inpxml):
     Test of the functionality of xml_set_attrib_value_no_create with multiple occurrences
     of the attribute and different values for occurrences
     """
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_attrib_value
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -255,7 +255,7 @@ def test_xml_set_attrib_value_differing_xpaths(load_inpxml):
 
 
 def test_xml_set_attrib_value_create(load_inpxml):
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_attrib_value
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -281,7 +281,7 @@ def test_xml_set_first_attrib_value(load_inpxml):
     Test of the functionality of xml_set_attrib_value_no_create with multiple occurrences
     of the sttribute and different values for occurrences
     """
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_first_attrib_value
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -296,7 +296,7 @@ def test_xml_set_first_attrib_value(load_inpxml):
 
 
 def test_xml_set_first_attrib_value_create(load_inpxml):
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_first_attrib_value
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -319,7 +319,7 @@ def test_xml_set_first_attrib_value_create(load_inpxml):
 
 def test_xml_set_text(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_text
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -343,7 +343,7 @@ TEST_TEXT_OCCURENCES = [None, None, None, [-1]]
 @pytest.mark.parametrize('text, expected_result,occurrences', zip(TEST_TEXTS, TEST_TEXT_RESULTS, TEST_TEXT_OCCURENCES))
 def test_xml_set_text_all(load_inpxml, text, expected_result, occurrences):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_text
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -362,7 +362,7 @@ def test_xml_set_text_all(load_inpxml, text, expected_result, occurrences):
 
 def test_xml_set_text_differing_xpaths(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_text
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -381,7 +381,7 @@ def test_xml_set_text_differing_xpaths(load_inpxml):
 
 def test_xml_set_text_create(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_text
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -403,7 +403,7 @@ def test_xml_set_text_create(load_inpxml):
 
 def test_xml_set_first_text(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_first_text
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -424,7 +424,7 @@ def test_xml_set_first_text(load_inpxml):
 
 def test_xml_set_first_text_create(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_first_text
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -455,7 +455,7 @@ TEST_SHIFT_RESULTS = ['13.0000000000', '9.3210000000', '13.0000000000']
                          zip(TEST_SHIFT_ATTRIB_NAME, TEST_SHIFT_VALUES, TEST_SHIFT_RESULTS))
 def test_xml_add_number_to_attrib(load_inpxml, attribname, shift_value, result):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_add_number_to_attrib
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -479,7 +479,7 @@ TEST_SHIFT_OCCURRENCES = [None, None, 0]
                          zip(TEST_SHIFT_VALUES_ALL, TEST_SHIFT_RESULTS_ALL_ABS, TEST_SHIFT_OCCURRENCES))
 def test_xml_add_number_to_attrib_all_abs(load_inpxml, shift_value, result, occurrences):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_add_number_to_attrib
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -502,7 +502,7 @@ def test_xml_add_number_to_attrib_all_abs(load_inpxml, shift_value, result, occu
                          zip(TEST_SHIFT_VALUES_ALL, TEST_SHIFT_RESULTS_ALL_REL, TEST_SHIFT_OCCURRENCES))
 def test_xml_add_number_to_attrib_all_rel(load_inpxml, shift_value, result, occurrences):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_add_number_to_attrib
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -524,7 +524,7 @@ def test_xml_add_number_to_attrib_all_rel(load_inpxml, shift_value, result, occu
 
 def test_xml_add_number_to_attrib_differing_xpaths(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_add_number_to_attrib
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -540,7 +540,7 @@ def test_xml_add_number_to_attrib_differing_xpaths(load_inpxml):
 
 def test_xml_add_number_to_attrib_differing_int(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_add_number_to_attrib
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -560,7 +560,7 @@ def test_xml_add_number_to_attrib_differing_int(load_inpxml):
 
 def test_xml_add_number_to_first_attrib(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_add_number_to_first_attrib
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -576,7 +576,7 @@ def test_xml_add_number_to_first_attrib(load_inpxml):
 
 def test_xml_set_simple_tag_single(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_simple_tag
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -594,7 +594,7 @@ def test_xml_set_simple_tag_single(load_inpxml):
 
 def test_xml_set_simple_tag_multiple_dict(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_simple_tag
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -612,7 +612,7 @@ def test_xml_set_simple_tag_multiple_dict(load_inpxml):
 
 def test_xml_set_simple_tag_multiple_list(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_simple_tag
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -631,7 +631,7 @@ def test_xml_set_simple_tag_multiple_list(load_inpxml):
 
 def test_xml_set_simple_tag_differing_xpaths(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_simple_tag
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -651,7 +651,7 @@ def test_xml_set_simple_tag_differing_xpaths(load_inpxml):
 
 def test_xml_set_simple_tag_create_single(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_simple_tag
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -677,7 +677,7 @@ def test_xml_set_simple_tag_create_single(load_inpxml):
 
 def test_xml_set_simple_tag_create_multiple(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_simple_tag
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -704,7 +704,7 @@ def test_xml_set_simple_tag_create_multiple(load_inpxml):
 
 def test_xml_set_complex_tag(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_complex_tag
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -724,7 +724,7 @@ def test_xml_set_complex_tag(load_inpxml):
 
 def test_xml_set_complex_tag_differing_xpaths_recursive_complex_single(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_complex_tag
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
@@ -741,7 +741,7 @@ def test_xml_set_complex_tag_differing_xpaths_recursive_complex_single(load_inpx
 
 def test_xml_set_complex_tag_recursive_complex_multiple(load_inpxml):
 
-    from masci_tools.util.xml.common_xml_util import eval_xpath
+    from masci_tools.util.xml.common_functions import eval_xpath
     from masci_tools.util.xml.xml_setters_xpaths import xml_set_complex_tag
 
     xmltree, schema_dict = load_inpxml(TEST_INPXML_PATH)
